feat: Extend GraphQL via Hooks
